Add `Archivable` concern
This is a concern that is used in models which can be archived, meaning
they have an `archived_at` column. This is currently only the case for
batches, but will be extended to consent forms in the future.
